# shared/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NULL HORIZON API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/progress": {
      "get": {
        "tags": [
          "progress"
        ],
        "summary": "Get Progress",
        "operationId": "get_progress_v1_progress_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/progress/missions/{mission_id}": {
      "put": {
        "tags": [
          "progress"
        ],
        "summary": "Put Mission Progress",
        "operationId": "put_mission_progress_v1_progress_missions__mission_id__put",
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mission Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertMissionProgressRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionProgressItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/progress/sync": {
      "post": {
        "tags": [
          "progress"
        ],
        "summary": "Sync Progress",
        "operationId": "sync_progress_v1_progress_sync_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressSyncRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressSyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MissionProgressItem": {
        "properties": {
          "mission_id": {
            "type": "string",
            "title": "Mission Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "best_hint_level": {
            "type": "integer",
            "title": "Best Hint Level",
            "default": 0
          },
          "clearance_awarded": {
            "type": "integer",
            "title": "Clearance Awarded",
            "default": 0
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count",
            "default": 1
          },
          "mission_version": {
            "type": "string",
            "title": "Mission Version",
            "default": "1.0.0"
          }
        },
        "type": "object",
        "required": [
          "mission_id",
          "status"
        ],
        "title": "MissionProgressItem"
      },
      "ProgressResponse": {
        "properties": {
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "rank": {
            "type": "string",
            "title": "Rank",
            "default": "Emergency Operator"
          },
          "clearance_points": {
            "type": "integer",
            "title": "Clearance Points",
            "default": 0
          },
          "missions": {
            "items": {
              "$ref": "#/components/schemas/MissionProgressItem"
            },
            "type": "array",
            "title": "Missions",
            "default": []
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillMasteryItem"
            },
            "type": "array",
            "title": "Skills",
            "default": []
          },
          "rewards": {
            "items": {
              "$ref": "#/components/schemas/RewardItem"
            },
            "type": "array",
            "title": "Rewards",
            "default": []
          },
          "review_recommendations": {
            "items": {
              "$ref": "#/components/schemas/ReviewRecommendationItem"
            },
            "type": "array",
            "title": "Review Recommendations",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "profile_id"
        ],
        "title": "ProgressResponse"
      },
      "ProgressSyncRequest": {
        "properties": {
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "missions": {
            "items": {
              "$ref": "#/components/schemas/MissionProgressItem"
            },
            "type": "array",
            "title": "Missions",
            "default": []
          },
          "skill_evidence": {
            "items": {
              "$ref": "#/components/schemas/SkillEvidenceItem"
            },
            "type": "array",
            "title": "Skill Evidence",
            "default": []
          },
          "rewards": {
            "items": {
              "$ref": "#/components/schemas/RewardItem"
            },
            "type": "array",
            "title": "Rewards",
            "default": []
          }
        },
        "type": "object",
        "title": "ProgressSyncRequest"
      },
      "ProgressSyncResponse": {
        "properties": {
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "accepted": {
            "type": "integer",
            "title": "Accepted"
          },
          "rank": {
            "type": "string",
            "title": "Rank"
          },
          "clearance_points": {
            "type": "integer",
            "title": "Clearance Points"
          },
          "missions": {
            "items": {
              "$ref": "#/components/schemas/MissionProgressItem"
            },
            "type": "array",
            "title": "Missions"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillMasteryItem"
            },
            "type": "array",
            "title": "Skills"
          },
          "rewards": {
            "items": {
              "$ref": "#/components/schemas/RewardItem"
            },
            "type": "array",
            "title": "Rewards"
          },
          "review_recommendations": {
            "items": {
              "$ref": "#/components/schemas/ReviewRecommendationItem"
            },
            "type": "array",
            "title": "Review Recommendations",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "profile_id",
          "accepted",
          "rank",
          "clearance_points",
          "missions",
          "skills",
          "rewards"
        ],
        "title": "ProgressSyncResponse"
      },
      "ReviewRecommendationItem": {
        "properties": {
          "skill_id": {
            "type": "string",
            "title": "Skill Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "skill_id",
          "reason"
        ],
        "title": "ReviewRecommendationItem"
      },
      "RewardItem": {
        "properties": {
          "reward_id": {
            "type": "string",
            "title": "Reward Id"
          },
          "equipped": {
            "type": "boolean",
            "title": "Equipped",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "reward_id"
        ],
        "title": "RewardItem"
      },
      "SkillEvidenceItem": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "skill_id": {
            "type": "string",
            "title": "Skill Id"
          },
          "mission_id": {
            "type": "string",
            "title": "Mission Id"
          },
          "assisted": {
            "type": "boolean",
            "title": "Assisted",
            "default": false
          },
          "delta": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Delta",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "skill_id",
          "mission_id"
        ],
        "title": "SkillEvidenceItem"
      },
      "SkillMasteryItem": {
        "properties": {
          "skill_id": {
            "type": "string",
            "title": "Skill Id"
          },
          "mastery_level": {
            "type": "string",
            "title": "Mastery Level"
          },
          "evidence_count": {
            "type": "integer",
            "title": "Evidence Count"
          },
          "unassisted_evidence_count": {
            "type": "integer",
            "title": "Unassisted Evidence Count"
          },
          "last_practiced_at_epoch_ms": {
            "type": "integer",
            "title": "Last Practiced At Epoch Ms",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "skill_id",
          "mastery_level",
          "evidence_count",
          "unassisted_evidence_count"
        ],
        "title": "SkillMasteryItem"
      },
      "UpsertMissionProgressRequest": {
        "properties": {
          "status": {
            "type": "string",
            "const": "completed",
            "title": "Status",
            "default": "completed"
          },
          "best_hint_level": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Best Hint Level",
            "default": 0
          },
          "clearance_awarded": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Clearance Awarded",
            "default": 0
          },
          "attempt_count": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Attempt Count",
            "default": 1
          },
          "mission_version": {
            "type": "string",
            "title": "Mission Version",
            "default": "1.0.0"
          }
        },
        "type": "object",
        "title": "UpsertMissionProgressRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}